Update openai.py - To Support Embedding in LM Studio
LM studio provide, by default, nomic-ai/nomic-embed-text-v1.5-GGUF for embedding.
When calling the exposed API, with tiktoken_enabled=True, the client sends integers as embedding requests.
The API expects Strings and it returns errors: [ERROR] 'input' field must be a string or an array of strings

With this update, users can set as False tiktoken_enabled when they need it. By default, the value is set to True (so, not changing current behaviour).